+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 - [Users having "Team" plan or above may now access shared directories in Cloud
   File Browser][12208]
 - [Added support for rendering numbered and nested lists][12190].
+- [Added buttons for editing top-level markdown elements in the documentation
+  panel][12217].
 - [Removed `#` from default colum name][12222]
 
 [11889]: https://github.com/enso-org/enso/pull/11889
@@ -34,6 +36,7 @@
 [12208]: https://github.com/enso-org/enso/pull/12208
 [12190]: https://github.com/enso-org/enso/pull/12190
 [12222]: https://github.com/enso-org/enso/pull/12222
+[12217]: https://github.com/enso-org/enso/pull/12217
 
 #### Enso Standard Library